Tuning.md
This project focuses on developing a portable and efficient mobile ventilator system designed to assist patients with respiratory difficulties, especially in scenarios where traditional ventilators are unavailable or impractical. This project is developed under the Faculty of Computer and Information Science at Ain Shams University in the field of Bioinformatics and holds copyrights from the faculty.